Implement return for definemethod.m4
Return in m4 methods are implemented by taking advantage of quoting.
return() expands into a start quote which treats makes m4 treat the
rest of the method as a string. At the method end, there is a dangling
end quote waiting. There is a default return() at the end of each
method which gets skipped over if there is an eariler return call.

One issue with this is that the dangling end quote causes issues if
the method is accessed with defn(). The workaround is to save each
method in a macro named using a unique id then define another macro
with the definition name which calls the method using indir.

An alternative workaround would be to just wrap the method inside
format() but the current workaround is superior in performance,
especially if using namespaces.